refactor(claude_statusline): apply domain-driven structure to models and layout configuration
- Dismantled `models.py` into distinct domains (`payload.py`, `layout.py`).
- Integrated `GitInfo` directly into `segments/git.py`.
- Integrated `CachedSegment` directly into `cache.py`.
- Eliminated previous arbitrary `types/` abstraction wrapper.
- Improved rendering strategy by compiling independent layout `Table.grid()` instances dynamically partitioned by line index, to mitigate shared-column width stretching.
