+++ src/python/claude_statusline/claude_statusline/payload.py
