--- src/python/claude_statusline/claude_statusline/types/cache.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-from whenever import Instant
-
-from claude_statusline.types.layout import SegmentGenerationResult
-
-
-class CachedSegment(BaseModel):
-    results: list[SegmentGenerationResult]
-    expires_at: Instant
