+++ src/python/claude_statusline/tests/test_segments_claude.py
@@ -1,5 +1,13 @@
 import unittest
 
+from claude_statusline.payload import (
+    AgentInfo,
+    ContextWindowInfo,
+    CostInfo,
+    ModelInfo,
+    OutputStyle,
+    StatusLineStdIn,
+)
 from claude_statusline.segments.claude import (
     format_context_usage,
     format_cost,
@@ -13,14 +21,6 @@ from claude_statusline.segments.constants import (
     ORANGE,
     RED,
     YELLOW,
-)
-from claude_statusline.types.payload import (
-    AgentInfo,
-    ContextWindowInfo,
-    CostInfo,
-    ModelInfo,
-    OutputStyle,
-    StatusLineStdIn,
 )
