+++ src/python/claude_statusline/tests/test_segments_git.py
@@ -11,8 +11,7 @@ from claude_statusline.segments.constants import (
     YELLOW,
     get_icon,
 )
-from claude_statusline.segments.git import format_git_full, generate_git_segment
-from claude_statusline.types.git import GitInfo
+from claude_statusline.segments.git import GitInfo, format_git_full, generate_git_segment
 
 
 def _git_info(**overrides: Any) -> GitInfo:
